+++ bricks/visual-builder/src/bootstrap.ts
@@ -17,3 +17,4 @@ import "./pre-generated-container/index.js";
 import "./property-editor/index.js";
 import "./data-providers/check-editor-by-name.js";
 import "./data-providers/get-data-dependency.js";
+import "./data-providers/get-dependency-tree.js";