+++ bricks/advanced/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.33.9](https://github.com/easyops-cn/next-bricks/compare/@next-bricks/advanced@0.33.8...@next-bricks/advanced@0.33.9) (2024-06-26)
+
+**Note:** Version bump only for package @next-bricks/advanced
+
+
+
+
+
+## [0.33.8](https://github.com/easyops-cn/next-bricks/compare/@next-bricks/advanced@0.33.7...@next-bricks/advanced@0.33.8) (2024-06-25)
+
+**Note:** Version bump only for package @next-bricks/advanced
+
+
+
+
+
 ## [0.33.7](https://github.com/easyops-cn/next-bricks/compare/@next-bricks/advanced@0.33.6...@next-bricks/advanced@0.33.7) (2024-06-25)
 
 **Note:** Version bump only for package @next-bricks/advanced