+++ bricks/visual-builder/CHANGELOG.md
@@ -3,6 +3,52 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.24.1](https://github.com/easyops-cn/next-bricks/compare/@next-bricks/visual-builder@1.24.0...@next-bricks/visual-builder@1.24.1) (2024-06-26)
+
+
+### Bug Fixes
+
+* code Editor output and update Editor refresh values ([9f33078](https://github.com/easyops-cn/next-bricks/commit/9f330787f9b9dd742f3970d8f6ef6bfe0ef2056a))
+
+
+
+
+
+# [1.24.0](https://github.com/easyops-cn/next-bricks/compare/@next-bricks/visual-builder@1.23.0...@next-bricks/visual-builder@1.24.0) (2024-06-26)
+
+
+### Bug Fixes
+
+* update color match RegExp ([9b44d57](https://github.com/easyops-cn/next-bricks/commit/9b44d578529ee4909683f9fc9e19bfd17b3b3e66))
+
+
+### Features
+
+* support colorPickComponent  and fix format ([2195a0d](https://github.com/easyops-cn/next-bricks/commit/2195a0deb7e75422ee2cd772df6e4b7c76c1642e))
+
+
+
+
+
+# [1.23.0](https://github.com/easyops-cn/next-bricks/compare/@next-bricks/visual-builder@1.22.2...@next-bricks/visual-builder@1.23.0) (2024-06-26)
+
+
+### Features
+
+* data dependency provider ([a55ec2a](https://github.com/easyops-cn/next-bricks/commit/a55ec2a606f53cff6f293304d2058e53e9312843))
+
+
+
+
+
+## [1.22.2](https://github.com/easyops-cn/next-bricks/compare/@next-bricks/visual-builder@1.22.1...@next-bricks/visual-builder@1.22.2) (2024-06-25)
+
+**Note:** Version bump only for package @next-bricks/visual-builder
+
+
+
+
+
 ## [1.22.1](https://github.com/easyops-cn/next-bricks/compare/@next-bricks/visual-builder@1.22.0...@next-bricks/visual-builder@1.22.1) (2024-06-25)
 
 **Note:** Version bump only for package @next-bricks/visual-builder