+++ shared/datetime/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.30](https://github.com/easyops-cn/next-bricks/compare/@next-shared/datetime@0.1.29...@next-shared/datetime@0.1.30) (2024-06-25)
+
+**Note:** Version bump only for package @next-shared/datetime
+
+
+
+
+
 ## [0.1.29](https://github.com/easyops-cn/next-bricks/compare/@next-shared/datetime@0.1.28...@next-shared/datetime@0.1.29) (2024-06-25)
 
 **Note:** Version bump only for package @next-shared/datetime