+++ shared/general/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.38](https://github.com/easyops-cn/next-bricks/compare/@next-shared/general@0.8.37...@next-shared/general@0.8.38) (2024-06-26)
+
+**Note:** Version bump only for package @next-shared/general
+
+
+
+
+
+## [0.8.37](https://github.com/easyops-cn/next-bricks/compare/@next-shared/general@0.8.36...@next-shared/general@0.8.37) (2024-06-25)
+
+**Note:** Version bump only for package @next-shared/general
+
+
+
+
+
 ## [0.8.36](https://github.com/easyops-cn/next-bricks/compare/@next-shared/general@0.8.35...@next-shared/general@0.8.36) (2024-06-25)
 
 **Note:** Version bump only for package @next-shared/general